Rewrite mod detection code
Completely rewrote how winterbot detects and responds to mods.  We no
longer rely on the op/deop messages in the chat room, instead we check
the mod list of the room on startup (and every 15 minutes, in case
someone gets removed, but I may have to add back the op/deop check for
"security").