Console changes
Change the WinterExtensions library into a console application, which
now loads the Winter specific options.
Give an error popup when the application isn't configured properly.
Add a bit more information to the console window.
Add an event for when the bot connects to a stream.
Minor change to options, now don't throw an exception when we don't have
a stream/user/oauth fields.